Allow user to request a raised exception if hash(content) will be inconsistent (#107)
* Update hashing.py

Allows user to have custom_hasher raise ValueError if it cannot identify a consistent-across-instantiations hash for the object.

This is valuable because we've found it very easy to create artifacts that are impossible to get reused, because they have data in their config that can’t be represented as Python primitives. Currently you can’t possibly know if you’ve created such an artifact until you try to rerun and notice it’s taking a long time / the reuse is not being picked up, or unless you look at the db. This change supports users in reducing the cycle time to identify this error.

* whitespace linting

* formatting